Removed CommonBotAdapter; moved BotOptions
