--- Source/Microsoft.Teams.Apps.CompanyCommunicator.Common/Services/CommonBot/CommonBotAdapter.cs
@@ -1,24 +1,0 @@
-﻿// <copyright file="CommonBotAdapter.cs" company="Microsoft">
-// Copyright (c) Microsoft. All rights reserved.
-// </copyright>
-
-namespace Microsoft.Teams.Apps.CompanyCommunicator.Common.Services.CommonBot
-{
-    using Microsoft.Bot.Builder.Integration.AspNet.Core;
-    using Microsoft.Bot.Connector.Authentication;
-
-    /// <summary>
-    /// A common bot adapter.
-    /// </summary>
-    public class CommonBotAdapter : BotFrameworkHttpAdapter
-    {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="CommonBotAdapter"/> class.
-        /// </summary>
-        /// <param name="credentialProvider">Credential provider service instance.</param>
-        public CommonBotAdapter(ICredentialProvider credentialProvider)
-            : base(credentialProvider)
-        {
-        }
-    }
-}
